+++ example/main.c
@@ -21,6 +21,101 @@ static const RGBColor_t rgb_sequence[3] =
 static const RGBColor_t rainbow_sequence[7] =
     { RED, ORANGE, YELLOW, GREEN, BLUE, INDIGO, VIOLET };
 
+/* Series of Patterns */
+/* This series starts with a RGB flash, slowly decreases the blank time until
+ * it hits a ribbon pattern, which slowly grows the band width. The series
+ * finishes by reversing - shrinking the ribbons and increasing the blank time.
+ */
+uint8_t rgb_spread_series_steps(void) {
+    return 10 + 20 + 10;
+}
+GenericPattern_t rgb_spread_series_patterns(void) {
+    static FlashArgs_t flash_args =
+        { .sequence = rgb_sequence, .length = 3, .color_delay = 200, .blank_delay = 200 };
+    static RibbonArgs_t ribbon_args =
+        { .sequence = rgb_sequence, .length = 3, .delay = 200 };
+    if (current_series_step < 10) {
+        flash_args.blank_delay = 200 - (20 * current_series_step);
+        GenericPattern_t current_pattern = FLASH_PATTERN(flash_args);
+        return current_pattern;
+    } else if (current_series_step < 20) {
+        ribbon_args.delay = 200 + 20 * (current_series_step - 10);
+        GenericPattern_t current_pattern = RIBBON_PATTERN(ribbon_args);
+        return current_pattern;
+    } else if (current_series_step < 30) {
+        ribbon_args.delay = 400 - 20 * (current_series_step - 20);
+        GenericPattern_t current_pattern = RIBBON_PATTERN(ribbon_args);
+        return current_pattern;
+    } else {
+        flash_args.blank_delay = 20 * (current_series_step - 30);
+        GenericPattern_t current_pattern = FLASH_PATTERN(flash_args);
+        return current_pattern;
+    }
+}
+static const SeriesArgs_t rgb_spread_series =
+    { .total_series_steps_function = rgb_spread_series_steps, .get_pattern_for_step = rgb_spread_series_patterns };
+/* This series starts as an RGB wide scroll, slowly adding colors until it
+ * becomes a Rainbow wide scroll, then slowly removing colors until it becomes
+ * RGB again.
+ */
+static RGBColor_t rgb_rainbow_sequence[7];
+uint8_t rgb_rainbow_scroll_steps(void) {
+    return 8;
+}
+GenericPattern_t rgb_rainbow_scroll_patterns(void) {
+    static WideScrollArgs_t args =
+        { .delay = 120, .sequence = rgb_rainbow_sequence, .length = 7 };
+    switch (current_series_step) {
+        case 0:
+            rgb_rainbow_sequence[0] = (RGBColor_t) RED;
+            rgb_rainbow_sequence[1] = (RGBColor_t) GREEN;
+            rgb_rainbow_sequence[2] = (RGBColor_t) BLUE;
+            args.length = 3;
+            break;
+        case 1: case 7:{
+            rgb_rainbow_sequence[0] = (RGBColor_t) RED;
+            rgb_rainbow_sequence[1] = (RGBColor_t) GREEN;
+            rgb_rainbow_sequence[2] = (RGBColor_t) BLUE;
+            rgb_rainbow_sequence[3] = (RGBColor_t) VIOLET;
+            args.length = 4;
+            break;
+        }
+        case 2: case 6: {
+            rgb_rainbow_sequence[0] = (RGBColor_t) RED;
+            rgb_rainbow_sequence[1] = (RGBColor_t) YELLOW;
+            rgb_rainbow_sequence[2] = (RGBColor_t) GREEN;
+            rgb_rainbow_sequence[3] = (RGBColor_t) BLUE;
+            rgb_rainbow_sequence[4] = (RGBColor_t) VIOLET;
+            args.length = 5;
+            break;
+        }
+        case 3: case 5: {
+            rgb_rainbow_sequence[0] = (RGBColor_t) RED;
+            rgb_rainbow_sequence[1] = (RGBColor_t) YELLOW;
+            rgb_rainbow_sequence[2] = (RGBColor_t) GREEN;
+            rgb_rainbow_sequence[3] = (RGBColor_t) BLUE;
+            rgb_rainbow_sequence[4] = (RGBColor_t) INDIGO;
+            rgb_rainbow_sequence[5] = (RGBColor_t) VIOLET;
+            args.length = 6;
+            break;
+        }
+        case 4: default:
+            rgb_rainbow_sequence[0] = (RGBColor_t) RED;
+            rgb_rainbow_sequence[1] = (RGBColor_t) ORANGE;
+            rgb_rainbow_sequence[2] = (RGBColor_t) YELLOW;
+            rgb_rainbow_sequence[3] = (RGBColor_t) GREEN;
+            rgb_rainbow_sequence[4] = (RGBColor_t) BLUE;
+            rgb_rainbow_sequence[5] = (RGBColor_t) INDIGO;
+            rgb_rainbow_sequence[6] = (RGBColor_t) VIOLET;
+            args.length = 7;
+            break;
+    }
+    GenericPattern_t pattern = WIDE_SCROLL_PATTERN(args);
+    return pattern;
+}
+static const SeriesArgs_t rgb_rainbow_scroll_series =
+    { .total_series_steps_function = rgb_rainbow_scroll_steps, .get_pattern_for_step = rgb_rainbow_scroll_patterns };
+
 
 int main(void) {
     apa102_init_spi();
@@ -48,6 +143,8 @@ int main(void) {
         FLASH_PATTERN(rainbow_flash),
         SCROLL_PATTERN(rgb_scroll),
         RIBBON_PATTERN(rainbow_ribbon),
+        SERIES_PATTERN(rgb_spread_series),
+        SERIES_PATTERN(rgb_rainbow_scroll_series),
     };
     const uint8_t pattern_count = (sizeof(patterns)) / (sizeof(patterns[0]));
 
@@ -78,9 +175,18 @@ int main(void) {
                 initialize_pattern(patterns + pattern_index);
             }
             for (uint8_t i = 0; i < 5; i++) {
-                run_step(patterns + pattern_index);
-                if (total_pattern_steps > 1) {
-                    while (current_pattern_step != 0) {
+                if (playing_pattern_series) {
+                    // Play through the first pattern
+                    for (uint8_t s = current_pattern_step; s < total_pattern_steps; s++) {
+                        run_step(patterns + pattern_index);
+                    }
+                    // Play through until the series is complete
+                    while (current_series_step != 0) {
+                        run_step(patterns + pattern_index);
+                    }
+                } else if (total_pattern_steps > 1) {
+                    // Play the pattern
+                    for (uint8_t s = current_pattern_step; s < total_pattern_steps; s++) {
                         run_step(patterns + pattern_index);
                     }
                 } else {